move tests unconsistent with api tests
