+++ test/lib/heuristics/scheduling_instance_validity_test.rb
@@ -22,7 +22,7 @@ class InstanceValidityTest < Minitest::Test
     def test_reject_if_no_heuristic_neither_first_sol_strategy
       problem = VRP.scheduling
       problem[:configuration][:preprocessing][:first_solution_strategy] = []
-      problem[:configuration][:resolution][:solver_parameter] = -1
+      problem[:configuration][:resolution][:solver] = false
 
       assert_includes OptimizerWrapper.config[:services][:ortools].inapplicable_solve?(TestHelper.create(problem)), :assert_solver_if_not_periodic
     end
@@ -127,25 +127,6 @@ class InstanceValidityTest < Minitest::Test
       refute_includes OptimizerWrapper.config[:services][:ortools].inapplicable_solve?(TestHelper.create(problem)), :assert_route_date_or_indice_if_periodic
     end
 
-    def test_assert_missions_in_route_exist
-      problem = VRP.scheduling
-      problem[:routes] = [{
-        vehicle_id: 'vehicle_0',
-        indice: 0,
-        mission_ids: ['service_1', 'service_3']
-      }]
-      refute_includes OptimizerWrapper.config[:services][:ortools].inapplicable_solve?(TestHelper.create(problem)), :assert_missions_in_routes_do_exist
-
-      problem[:routes] = [{
-        vehicle_id: 'vehicle_0',
-        indice: 0,
-        mission_ids: ['service_111', 'service_3']
-      }]
-      assert_raises OptimizerWrapper::DiscordantProblemError do
-        TestHelper.create(problem)
-      end
-    end
-
     def test_not_too_many_visits_provided_in_route
       problem = VRP.scheduling
       problem[:routes] = [{
@@ -174,32 +155,6 @@ class InstanceValidityTest < Minitest::Test
 
       problem[:configuration][:preprocessing][:first_solution_strategy] = []
       assert_includes OptimizerWrapper.config[:services][:ortools].inapplicable_solve?(TestHelper.create(problem)), :assert_no_route_if_schedule_without_periodic_heuristic
-    end
-
-    def test_same_point_day_authorized
-      vrp = VRP.scheduling
-      reference_point = vrp[:services].first[:activity][:point_id]
-      vrp[:services].first[:visits_number] = 3
-      vrp[:services].first[:minimum_lapse] = 3
-      vrp[:services].first[:maximum_lapse] = 3
-      vrp[:services] << {
-        id: 'last_service',
-        visits_number: 2,
-        minimum_lapse: 6,
-        maximum_lapse: 6,
-        activity: {
-          point_id: reference_point
-        }
-      }
-      vrp[:configuration][:resolution][:same_point_day] = true
-      vrp[:configuration][:schedule][:range_indices][:end] = 10
-      result = OptimizerWrapper.wrapper_vrp('demo', { services: { vrp: [:demo] }}, TestHelper.create(vrp), nil)
-      assert result # there exist a common_divisor
-
-      vrp[:services].last[:minimum_lapse] = vrp[:services].last[:maximum_lapse] = 7
-      assert_raises OptimizerWrapper::UnsupportedProblemError do
-        OptimizerWrapper.wrapper_vrp('demo', { services: { vrp: [:demo] }}, TestHelper.create(vrp), nil)
-      end
     end
   end
 end
