+++ test/lib/interpreters/multi_trips_test.rb
@@ -20,65 +20,10 @@ require 'date'
 
 class MultiTripsTest < Minitest::Test
   def test_expand_vehicles_trips
-    size = 5
-    problem = {
-      matrices: [{
-        id: 'matrix_0',
-        time: [
-          [0,  1,  1, 10,  0],
-          [1,  0,  1, 10,  1],
-          [1,  1,  0, 10,  1],
-          [10, 10, 10, 0, 10],
-          [0,  1,  1, 10,  0]
-        ],
-        distance: [
-          [0,  1,  1, 10,  0],
-          [1,  0,  1, 10,  1],
-          [1,  1,  0, 10,  1],
-          [10, 10, 10, 0, 10],
-          [0,  1,  1, 10,  0]
-        ]
-      }],
-      points: (0..(size - 1)).collect{ |i|
-        {
-          id: "point_#{i}",
-          matrix_index: i
-        }
-      },
-      vehicles: [{
-        id: 'vehicle_0',
-        start_point_id: 'point_0',
-        matrix_id: 'matrix_0',
-        timewindow: {
-          start: 1,
-          end: 2
-        },
-        trips: 2
-      }],
-      services: (1..(size - 1)).collect{ |i|
-        {
-          id: "service_#{i}",
-          activity: {
-            point_id: "point_#{i}",
-            timewindows: [{
-              start: 1,
-              end: 2
-            }, {
-              start: 5,
-              end: 7
-            }]
-          }
-        }
-      },
-      configuration: {
-        resolution: {
-          duration: 10
-        }
-      }
-    }
-    vrp = TestHelper.create(problem)
-    periodic = Interpreters::MultiTrips.new
-    periodic.send(:expand, vrp)
+    vrp = VRP.lat_lon
+    vrp[:vehicles].first[:trips] = 2
+
+    vrp = Interpreters::MultiTrips.new.expand(TestHelper.create(vrp))
 
     assert_equal 2, vrp.vehicles.size
     assert_equal 1, vrp.relations.size
@@ -89,7 +34,19 @@ class MultiTripsTest < Minitest::Test
     }
   end
 
+  def test_consecutive_expand
+    vrp = VRP.lat_lon_two_vehicles
+    vrp[:vehicles].first[:trips] = 2
+
+    vrp = TestHelper.create(vrp)
+    Interpreters::MultiTrips.new.expand(vrp)
+    assert_equal 3, vrp.vehicles.size
+    Interpreters::MultiTrips.new.expand(vrp) # consecutive MultiTrips.expand should not produce any error
+    assert_equal 3, vrp.vehicles.size
+  end
+
   def test_solve_vehicles_trips
+    # this test will be cleaned and moved with multi_tour developments
     size = 5
     problem = {
       units: [{
